# docs/swagger.yaml
basePath: /
definitions:
  helpers.CommonErrorResponse:
    properties:
      error:
        example: Bad Request
        type: string
      message: {}
      statusCode:
        example: 400
        type: integer
    type: object
  helpers.CommonResponse:
    properties:
      data: {}
      message:
        example: Success
        type: string
      meta:
        $ref: '#/definitions/helpers.Meta'
      statusCode:
        example: 200
        type: integer
    type: object
...
  helpers.Meta:
    properties:
      currentPage:
        example: 1
        type: integer
      from:
        example: 1
        type: integer
      lastPage:
        example: 5
        type: integer
      limit:
        example: 10
        type: integer
      to:
        example: 10
        type: integer
      total:
        example: 42
        type: integer
    type: object
...
  models.Authors:
    properties:
      bio:
        type: string
      books:
        items:
          $ref: '#/definitions/models.Books'
        type: array
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      nationality:
        type: string
      updated_at:
        type: string
    type: object
  models.Books:
    properties:
      author:
        $ref: '#/definitions/models.Authors'
      author_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      genre:
        type: string
      id:
        type: integer
      isbn:
        type: string
      publisher:
        $ref: '#/definitions/models.Publishers'
      publisher_id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      year:
        type: integer
    type: object
  models.Publishers:
    properties:
      address:
        type: string
      books:
        items:
          $ref: '#/definitions/models.Books'
        type: array
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      website:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@pubcore.io
    name: API Support
  description: Publishing platform API for managing Books, Authors, and Publishers.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PubCore API
  version: "1.0"
paths:
...
  /authors:
    get:
      description: Retrieve a paginated, filterable, sortable list of authors
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - default: created_at
        description: Sort field (e.g. name, created_at)
        in: query
        name: sort
        type: string
      - default: desc
        description: Sort order (asc or desc)
        in: query
        name: order
        type: string
      - description: Filter by name (partial match)
        in: query
        name: name
        type: string
      - description: Filter by nationality
        in: query
        name: nationality
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.CommonResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.CommonErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.CommonErrorResponse'
      security:
      - BearerAuth: []
      summary: List all authors
      tags:
      - Authors
    post:
      consumes:
      - application/json
      description: Add a new author to the platform
      parameters:
      - description: Author Input
        in: body
        name: author
        required: true
        schema:
          $ref: '#/definitions/models.Authors'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helpers.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.CommonErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new author
      tags:
      - Authors
  /authors/{id}:
    delete:
      description: Remove an author by ID (only if they have no books)
      parameters:
      - description: Author ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.CommonErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.CommonErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete an author
      tags:
      - Authors
    get:
      description: Retrieve an author and their books by ID
      parameters:
      - description: Author ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Authors'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a single author by ID
      tags:
      - Authors
    put:
      consumes:
      - application/json
      description: Modify author details by ID
      parameters:
      - description: Author ID
        in: path
        name: id
        required: true
        type: integer
      - description: Author Update Input
        in: body
        name: author
        required: true
        schema:
          $ref: '#/definitions/models.Authors'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Authors'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update an existing author
      tags:
      - Authors
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"